% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cj_apns.R
\name{.flatten_estimates}
\alias{.flatten_estimates}
\title{Flatten point estimates into a named vector}
\usage{
.flatten_estimates(pt, estimand, assumption, attr_names)
}
\description{
Flatten point estimates into a named vector
}
\keyword{internal}
