% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{.folded_normal_mean}
\alias{.folded_normal_mean}
\title{Mean of a folded normal distribution}
\usage{
.folded_normal_mean(mu, sigma)
}
\description{
Mean of a folded normal distribution
}
\keyword{internal}
