% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomization.R
\name{.ri_warn_checks}
\alias{.ri_warn_checks}
\title{Emit consolidated warnings from the design checks}
\usage{
.ri_warn_checks(check_store)
}
\description{
Emit consolidated warnings from the design checks
}
\keyword{internal}
