% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{.sim_cond_amce}
\alias{.sim_cond_amce}
\title{Simulate a conditional AMCE from the parametric bootstrap}
\usage{
.sim_cond_amce(cond_amce_info, a, tq, tp, base, grp)
}
\description{
Simulate a conditional AMCE from the parametric bootstrap
}
\keyword{internal}
