% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.cj_apns}
\alias{plot.cj_apns}
\title{Plot method for cj_apns objects}
\usage{
\method{plot}{cj_apns}(x, what = NULL, ...)
}
\arguments{
\item{x}{An object of class `"cj_apns"`.}

\item{what}{Which estimand to plot: `"eapns"`, `"epns"`, or `"amce"`.
Defaults to the estimand stored in `x`.}

\item{...}{Passed to ggplot2 functions.}
}
\value{
A ggplot2 object (invisibly).
}
\description{
Produces a dot-and-whisker plot of EAPNS, EPNS, or AMCE estimates,
replicating the style of Figure 2 in Stoetzer & Magazinnik (2026).
Requires \pkg{ggplot2}.
}
\examples{
\dontrun{
data(cnj_cand)
res <- cj_apns(vote ~ borders + eurobonds + immucard + schools,
               data = cnj_cand, id = ~ ResponseId,
               se = "parametric", B = 200)
plot(res)
}

}
