% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.cj_apns}
\alias{print.cj_apns}
\title{Print method for cj_apns objects}
\usage{
\method{print}{cj_apns}(x, ...)
}
\arguments{
\item{x}{An object of class `"cj_apns"`.}

\item{...}{Ignored.}
}
\description{
Print method for cj_apns objects
}
