% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.cj_apns}
\alias{summary.cj_apns}
\title{Summary method for cj_apns objects}
\usage{
\method{summary}{cj_apns}(object, ...)
}
\arguments{
\item{object}{An object of class `"cj_apns"`.}

\item{...}{Ignored.}
}
\description{
Prints a detailed summary including pairwise EPNS and conditional AMCEs.
}
