% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amce.R
\name{.cluster_se_dim}
\alias{.cluster_se_dim}
\title{Cluster-robust SE for a difference-in-means}
\usage{
.cluster_se_dim(Y, treatment, level_tq, level_tp, cluster)
}
\description{
Cluster-robust SE for a difference-in-means
}
\keyword{internal}
